requirements.txt
streamlit
google-generativeai
langchain
langchain-google-genai
deep-translator
textwrap3
python-dotenv